\name{getMapAus}
\alias{getMapAus}
\title{Load Australian maps}
\usage{
  getMapAus(type = "SD", crop = TRUE)
}
\arguments{
  \item{type}{a map type}

  \item{crop}{a logical value determining whether map
  should be cropped to mainland Australia and Tasmania
  (i.e. excluding Cocos Keeling Islands, Christmas Island
  and Lord Howe Island).}
}
\description{
  Load Australian maps
}

