% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rmapaus}
\alias{rmapaus}
\alias{rmapaus-package}
\title{rmapaus: mapping Australia}
\description{
The rmapaus package provides spatial boundaries for various Australian regions,
including postcodes and Australian Bureau of Statistics statistical regions.
}

