% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scale_fill_heat}
\alias{scale_fill_heat}
\title{Add guide colours}
\usage{
scale_fill_heat(labels = waiver(), ...)
}
\arguments{
\item{labels}{defaults to \code{waiver()}. Can be for example,
\code{percent}}

\item{...}{additional arguments passed to scale functions}
}
\description{
Add guide colours
}

