\docType{data}
\name{SD.full}
\alias{SD.full}
\title{Statistical division map data}
\format{A spatial polygon data frame with 61 regions}
\usage{
data(sd)
}
\description{
A dataset containing map data for Australian statistical
divisions (SD). Data is sourced from the Australian Bureau
of Statistics.
}
\keyword{datasets}

