% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{SD.smoothed}
\alias{SD.smoothed}
\title{Statistical division map data (smoothed)}
\format{A spatial polygon data frame with 61 regions}
\usage{
data(sd)
}
\description{
A dataset containing map data for Australian statistical divisions (SD).
The regions have been adapted from the full ABS specification by smoothing
the boundaries to reduce the size of the data set and speed up plotting.
}
\keyword{datasets}

