% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{SSD.full}
\alias{SSD.full}
\title{Statistical sub-division map data}
\format{A spatial polygon data frame with 206 regions}
\usage{
data(ssd)

data(ssd)
}
\description{
A dataset containing map data for Australian statistical sub-divisions (SSD).
Data is sourced from the Australian Bureau of Statistics.

A dataset containing map data for Australian statistical sub-divisions (SSD).
}
\keyword{datasets}

