% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{SSD.smoothed}
\alias{SSD.smoothed}
\title{Statistical sub-division map data (smoothed)}
\format{A spatial polygon data frame with 206 regions}
\usage{
data(ssd)
}
\description{
A dataset containing map data for Australian statistical sub-divisions (SSD).
The regions have been adapted from the full ABS specification by smoothing
the boundaries to reduce the size of the data set and speed up plotting.
}
\keyword{datasets}

