\docType{data}
\name{STE.caricature}
\alias{STE.caricature}
\title{State and Territory mapdata (caricature)}
\format{A spatial polygon data frame with 206 regions}
\usage{
data(ste)
}
\description{
A dataset containing map data for Australian states and
territories The regions have been adapted from the full ABS
specification by smoothing the boundaries to reduce the
size of the data set and speed up plotting. These regions
have a higher level of smoothing than the "smoothed" data
set.
}
\keyword{datasets}

