% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{STE.full}
\alias{STE.full}
\title{State and Territory map data}
\format{A spatial polygon data frame with 9 regions}
\usage{
data(ste)
}
\description{
A dataset containing map data for Australian states and territories.
}
\keyword{datasets}

