\docType{data}
\name{STE.smoothed}
\alias{STE.smoothed}
\title{State and Territory mapdata (smoothed)}
\format{A spatial polygon data frame with 9 regions}
\usage{
data(ste)
}
\description{
A dataset containing map data for Australian states and
territories The regions have been adapted from the full ABS
specification by smoothing the boundaries to reduce the
size of the data set and speed up plotting.
}
\keyword{datasets}

