\docType{data}
\name{POA.full}
\alias{POA.full}
\title{Australian postcode map data}
\format{A spatial polygon data frame with 2513 regions}
\description{
  A dataset containing map data for Australian postcodes
  (POA). Data is sourced from the Australian Bureau of
  Statistics.
}
\keyword{datasets}

