% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scale_colour_size}
\alias{scale_colour_size}
\title{Add guide size}
\usage{
scale_colour_size(colour = "black", ...)
}
\arguments{
\item{colour}{colour of size bubbles in legend. Default is \code{black}.}

\item{...}{additional arguments passed to scale functions}
}
\description{
Add guide size
}

