\docType{data}
\name{SD.smoothed}
\alias{SD.smoothed}
\title{Statistical division map data (smoothed)}
\format{A data frame with 53940 rows and 10 variables}
\description{
  A dataset containing map data for Australian statistical
  divisions (SD). The regions have been adapted from the
  full ABS specification by smoothing the boundaries to
  reduce the size of the data set and speed up plotting.
}
\keyword{datasets}

