\docType{data}
\name{SLA.full}
\alias{SLA.full}
\title{Statistical local area map data}
\format{A spatial polygon data frame with 1390 regions}
\usage{
data(sla)
}
\description{
A dataset containing map data for Australian statistical
local areas (SLA). Data is sourced from the Australian
Bureau of Statistics.
}
\keyword{datasets}

